+++ CHANGELOG.md
@@ -6,7 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+## Changed
+
 - **Table**: The column resizing algorithm has been rewritten. Columns can now push each other.
+- **Button**, **SingleSelect**, **MultiSelect**, **Textfield**, **Textarea**: Use 4px border radius.
 
 ## [2.4.0] - 2025-12-26